requirements.txt
numpy
pandas
PyYAML
python-dotenv
pymongo
scikit-learn
mlflow
python-multipart
ensure
matplotlib
seaborn
joblib
notebook
tqdm
types-pyYAML
dagshub
fastapi
uvicorn
pyaml
